+++ ForeignKey.php
@@ -86,13 +86,20 @@ class ForeignKey
     /**
      * Get columns name.
      *
-     * @param bool $quoted
+     * @param bool $quoted Deprecated, use Hector\Query\Statement\Quoted instead
      * @param string|null $tableAlias
      *
      * @return string[]
      */
     public function getColumnsName(bool $quoted = false, ?string $tableAlias = null): array
     {
+        if ($quoted) {
+            trigger_error(
+                'The $quoted parameter of ' . __METHOD__ . '() is deprecated, use Hector\Query\Statement\Quoted instead.',
+                E_USER_DEPRECATED
+            );
+        }
+
         $names = $this->columns_name;
 
         $quoted && $names = $this->quoteNames($names);
@@ -124,13 +131,20 @@ class ForeignKey
     /**
      * Get referenced columns name.
      *
-     * @param bool $quoted
+     * @param bool $quoted Deprecated, use Hector\Query\Statement\Quoted instead
      * @param string|null $tableAlias
      *
      * @return string[]
      */
     public function getReferencedColumnsName(bool $quoted = false, ?string $tableAlias = null): array
     {
+        if ($quoted) {
+            trigger_error(
+                'The $quoted parameter of ' . __METHOD__ . '() is deprecated, use Hector\Query\Statement\Quoted instead.',
+                E_USER_DEPRECATED
+            );
+        }
+
         $names = $this->referenced_columns_name;
 
         $quoted && $names = $this->quoteNames($names);